feat: add central error handler

+++ controllers/movies.js
@@ -1,11 +1,13 @@
 const Movie = require('../models/movie');
+const NotFoundError = require('../errors/NotFoundError');
+const ForbiddenError = require('../errors/ForbiddenError');
 
 module.exports.getMovies = (req, res, next) => {
   const userId = req.user._id;
   Movie.find({ owner: userId })
     .then((movies) => {
       if (!movies) {
-        // throw new NotFoundError('Фильмы не найдены');
+        throw new NotFoundError('Фильмы не найдены');
       }
       return res.status(200).send(movies);
     })
@@ -26,7 +28,6 @@ module.exports.createMovie = (req, res, next) => {
     thumbnail,
     movieId,
   } = req.body;
-  const owner = req.user._id;
   return Movie.create({
     country,
     director,
@@ -39,15 +40,9 @@ module.exports.createMovie = (req, res, next) => {
     nameEN,
     thumbnail,
     movieId,
-    owner,
+    owner: req.user._id,
   })
     .then((movie) => res.status(200).send(movie))
-    .catch((err) => {
-      res.send(err);
-      if (err.name === 'ValidationError') {
-        // throw new BadRequestError('Переданы некорректные данные');
-      }
-    })
     .catch(next);
 };
 
@@ -56,7 +51,10 @@ module.exports.deleteMovie = (req, res, next) => {
   Movie.findById(id)
     .then((movie) => {
       if (!movie) {
-        // throw new NotFoundError('Фильм не найден');
+        throw new NotFoundError('Фильм не найден');
+      }
+      if (movie.owner.toString() !== req.user._id) {
+        throw new ForbiddenError('Попытка удалить фильм другого пользователя');
       }
       Movie.findByIdAndRemove(id)
         .then((deletedMovie) => {
